requirements.txt
joblib
scikit-image
matplotlib
opencv-python
tqdm
pyyaml
tensorboardx
wandb
pycocotools
classy_vision
absl-py
asttokens
autogra
backcall
batchgenerators
cachetools
certifi
charset-normalizer
colorama
coloredlogs
comm
contourpy
cycler
debugpy
decorator
einops
executing
flatbuffers
fonttools
future
google-auth
google-auth-oauthlib
grpcio
hausdorff
humanfriendly
icecream
idna
imageio
importlib-metadata
importlib-resources
ipykernel
ipython
jedi
joblib
jupyter_client
jupyter_core
kiwisolver
lazy_loader
linecache2
llvmlite
Markdown
MarkupSafe
matplotlib
matplotlib-inline
monai
mpmath
nest-asyncio
networkx
numba
numpy
oauthlib
onnx
onnxruntime
opencv-python
packaging
pandas
parso
pexpect
pickleshare
Pillow
platformdirs
prompt-toolkit
protobuf
psutil
ptyprocess
pure-eval
pyasn1
pyasn1-modules
pycocotools
Pygments
pyparsing
python-dateutil
pytz
PyWavelets
pyzmq
requests
requests-oauthlib
rsa
scikit-image
scikit-learn
scipy
seaborn
six
stack-data
sympy
tensorboard
tensorboard-data-server
threadpoolctl
tifffile
torch==1.12.1+cu116
torchvision==0.13.1+cu116
tornado
traceback2
traitlets
typing_extensions
tzdata
unittest2
urllib3
wcwidth
Werkzeug
zipp